# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your API Title",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/home": {
            "get": {
                "tags": [
                    "Home"
                ],
                "summary": "Get home data for authenticated user",
                "description": "Fetch home data for a user, including property listings and payment status.",
                "operationId": "d65ef76cbb32ae304c5a24c525b95adb",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved home data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "location": {
                                                "type": "string"
                                            },
                                            "price": {
                                                "type": "integer"
                                            },
                                            "like": {
                                                "type": "integer"
                                            },
                                            "Fridge": {
                                                "type": "integer"
                                            },
                                            "WaterTank": {
                                                "type": "stringr"
                                            },
                                            "Solar": {
                                                "type": "string"
                                            },
                                            "Roommates": {
                                                "type": "string"
                                            },
                                            "wifi": {
                                                "type": "string"
                                            },
                                            "geo": {
                                                "type": "string"
                                            },
                                            "image": {
                                                "type": "string",
                                                "format": "url",
                                                "example": "http://yourdomain.com/storage/properties/room.jpg"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Payment required",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "redirect_url": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/payment/regpayment": {
            "post": {
                "tags": [
                    "RegPayments"
                ],
                "description": "Processes a payment through Paynow using mobile money.",
                "operationId": "d433ff36fff558355b9db22a9063cac2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone",
                                    "paymentmethod",
                                    "amount"
                                ],
                                "properties": {
                                    "phone": {
                                        "type": "string",
                                        "example": "0771234567"
                                    },
                                    "paymentmethod": {
                                        "type": "string",
                                        "example": "ecocash"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 10
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transaction successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Transaction successful!"
                                        },
                                        "reference_number": {
                                            "type": "string",
                                            "example": "PN1234567890"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Transaction failed or cancelled",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Transaction cancelled"
                                        },
                                        "action": {
                                            "type": "string",
                                            "example": "Please try again or contact support."
                                        },
                                        "redirect_url": {
                                            "type": "string",
                                            "example": "http://your-app.com/payment/retry"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Paynow service unavailable",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Service is down, please try again later."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/upload": {
            "post": {
                "tags": [
                    "Upload-user-profile"
                ],
                "summary": "Update user profile",
                "description": "Update user's phone number and profile image",
                "operationId": "b56ec7ecfe9eaa4524204d8346f7c223",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "image",
                                    "phone"
                                ],
                                "properties": {
                                    "image": {
                                        "description": "Profile image to upload",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "phone": {
                                        "description": "User phone number",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile updated successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/profile/update": {
            "post": {
                "tags": [
                    "Upload-user-profile"
                ],
                "summary": "Update user profile",
                "description": "Allows a logged-in user to update their profile information such as name, surname, email, university, phone, password, image, etc.",
                "operationId": "updateProfile",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "phone"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "surname": {
                                        "type": "string",
                                        "example": "Doe"
                                    },
                                    "image": {
                                        "type": "file"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+123456789"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Profile updated successfully."
                                        },
                                        "properties": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "redirect_url": {
                                            "type": "string",
                                            "example": "http://yourdomain.com/home"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Profile updated successfully. Payment required.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Profile updated successfully. Payment required."
                                        },
                                        "redirect_url": {
                                            "type": "string",
                                            "example": "http://yourdomain.com/payment/regpayment"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use Bearer token to access protected routes",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}